Add a visually hidden h2 heading to subnav
When we were adding the cookie banner, we noticed that because the cookie banner
uses a h2 heading, the subnav headings on pages like /patterns look like they are
nested beneath the cookie banner content to anyone navigating using a screenreader looking
just at headings.

This doesn't match the actual structure of the page and could be confusing to users.

This commit adds a visually hidden h2 heading above the subnav headings to:

1) provide some context to the user as to what the headings actually relate to
2) make sure the subnav headings could no longer be mistaken as being related to the cookie banner content